Gets rid of pkl files.
