+++ Backend/app/ai/topics.py
@@ -1,13 +1,8 @@
 import json
 import pickle
-import random
-from typing import Literal
 
-from app.ai.input_processor import InputProcessor
-from app.file_paths import TOPIC_TYPES_FILE, TOPICS_FILE, WORDS_FILE
-
-IN = 0
-OUT = 1
+from app.ai.input_cleaner import InputCleaner
+from app.file_paths import TOPICS_FILE
 
 training_pair = tuple[list[str], str]
 
@@ -17,7 +12,7 @@ class Topics:
     def __init__(self) -> None:
         self._topics_list = _load_topic_list()
 
-        self._input_processor = InputProcessor()
+        self._input_cleaner = InputCleaner()
 
         self._topic_types: list[str] = []
         self._words: list[str] = []
@@ -36,82 +31,16 @@ class Topics:
         self._cleanup_words()
 
     def _append_pattern(self, pattern: str, topic_type: str) -> None:
-        word_list = self._input_processor.cleanup_sentence(pattern)
+        word_list = self._input_cleaner.cleanup_sentence(pattern)
         self._words.extend(word_list)
 
         self._training_pairs.append((word_list, topic_type))
 
     def _cleanup_words(self):
-        words = [self._input_processor.cleanup_word(word) for word in self._words]
+        words = [self._input_cleaner.cleanup_word(word) for word in self._words]
         words = sorted(set(words))
 
         self._words = words
-
-    def save_words(self):
-        with open(TOPIC_TYPES_FILE, "wb") as types_file:
-            pickle.dump(self._topic_types, types_file)
-
-        with open(WORDS_FILE, "wb") as words_file:
-            pickle.dump(self._words, words_file)
-
-    def processed_training_data(
-        self,
-    ) -> tuple[
-        list[list[Literal[0] | Literal[1]]], list[list[Literal[0] | Literal[1]]]
-    ]:
-        """
-        Returns a tuple of input and output data.
-        For each element of the input data, its corresponding expected output will be on the other list at the same index.
-        """
-
-        training = []
-
-        for training_pair in self._training_pairs:
-            occurrence_list = self._create_word_occurrence_list(training_pair[IN])
-            output_row = self._create_output_row(training_pair[OUT])
-
-            training.append([occurrence_list, output_row])
-        random.shuffle(training)
-
-        inputs_list, outputs_list = (list(column) for column in zip(*training))
-
-        return inputs_list, outputs_list
-
-    def _create_word_occurrence_list(
-        self, word_patterns: list[str]
-    ) -> list[Literal[0] | Literal[1]]:
-        """
-        Creates a list of word occurrence for the given training pair.
-
-        The elements of the list will contain 1 if its `self._words` matching element appears in the
-        word patterns of the given training pair, otherwise it'll be 0.
-
-        This list can be used as input for the training of our AI.
-        """
-
-        occurrence_list = []
-
-        words = [self._input_processor.cleanup_word(word) for word in word_patterns]
-
-        for word in self._words:
-            if word in words:
-                occurrence_list.append(1)
-            else:
-                occurrence_list.append(0)
-
-        return occurrence_list
-
-    def _create_output_row(self, topic_type: str) -> list[Literal[0] | Literal[1]]:
-        """
-        Creates a list that represents the desired output that our AI will give for a given topic_type.
-        """
-        type_index = self._topic_types.index(topic_type)
-
-        output_row = [0] * len(self._topic_types)
-
-        output_row[type_index] = 1
-
-        return output_row
 
     def get_topic(self, index: int) -> str:
         return self._topic_types[index]
@@ -121,13 +50,18 @@ class Topics:
             if topic["type"] == topic_name:
                 return topic["responses"][0]
 
+    def get_topic_types(self):
+        return self._topic_types
+
+    def get_words(self):
+        return self._words
+
+    def get_training_pairs(self):
+        return self._training_pairs
+
 
 def _load_topic_list():
-    with open(TOPICS_FILE) as topics_document:
+    with open(TOPICS_FILE, encoding="utf-8") as topics_document:
         topic_list = json.loads(topics_document.read())
 
     return topic_list["topics"]
-
-
-if __name__ == "__main__":
-    print(Topics().processed_training_data()[1])
